+++ src/Libraries/Microsoft.Extensions.Diagnostics.Testing/Logging/FakeLogCollector.cs
@@ -4,6 +4,8 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
 using Microsoft.Shared.Diagnostics;
 
@@ -14,7 +16,7 @@ namespace Microsoft.Extensions.Logging.Testing;
 /// </summary>
 [DebuggerDisplay("Count = {Count}, LatestRecord = {LatestRecord}")]
 [DebuggerTypeProxy(typeof(FakeLogCollectorDebugView))]
-public class FakeLogCollector
+public partial class FakeLogCollector
 {
     private readonly List<FakeLogRecord> _records = [];
     private readonly FakeLogCollectorOptions _options;
@@ -53,8 +55,14 @@ public class FakeLogCollector
     {
         lock (_records)
         {
-            _records.Clear();
+            ClearRecordsCore();
         }
+    }
+
+    private void ClearRecordsCore()
+    {
+        _records.Clear();
+        _ = Interlocked.Increment(ref _recordCollectionVersion);
     }
 
     /// <summary>
@@ -71,7 +79,7 @@ public class FakeLogCollector
             var records = _records.ToArray();
             if (clear)
             {
-                _records.Clear();
+                ClearRecordsCore();
             }
 
             return records;
@@ -136,10 +144,22 @@ public class FakeLogCollector
             return;
         }
 
+        TaskCompletionSource<object?>? logEnumerationSharedWaiterToWake = null;
+
         lock (_records)
         {
             _records.Add(record);
+
+            if (_waitingEnumeratorCount > 0)
+            {
+                logEnumerationSharedWaiterToWake = _logEnumerationSharedWaiter;
+                _logEnumerationSharedWaiter = new TaskCompletionSource<object?>(TaskCreationOptions.RunContinuationsAsynchronously);
+                _waitingEnumeratorCount = 0;
+            }
         }
+
+        // it is possible the task was already completed, but it does not matter and we can avoid locking
+        _ = logEnumerationSharedWaiterToWake?.TrySetResult(null);
 
         _options.OutputSink?.Invoke(_options.OutputFormatter(record));
     }